containers: separate kubernetes provider mixin
This patch separates the kubernetes provider specifics from the
container provider mixin into KubernetesProviderMixin.
